requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
google-auth==2.23.4
httpx==0.25.1
python-dotenv==1.0.0